+++ megamek/src/megamek/common/Mech.java
@@ -761,7 +761,7 @@ public abstract class Mech
      * Returns the engine rating
      */
     public int engineRating() {
-        return (int)Math.round(walkMP * weight);
+        return Math.round(walkMP * weight);
     }
     
     /**
@@ -1890,7 +1890,7 @@ public abstract class Mech
             (hasC3S() && C3Master > NONE) ||
             (hasC3i() && calculateFreeC3Nodes() < 5) ||
             assumeLinkedC3) {
-                xbv = (double)(Math.round(0.35 * weaponsBVFront + (0.5 * weaponsBVRear)));
+                xbv = Math.round(0.35 * weaponsBVFront + (0.5 * weaponsBVRear));
         }
 
         // Possibly adjust for TAG and Arrow IV.
